Update index.tsx
When refactoring a large component, one of the first things I would look for are business logic and functions that are not directly tied to the component state. 
I have created small components to be able to have a smaller structure in which to organize our FrontEnd

- Small component
- Funcional Files
-Types
-Logical files

Now the function and the business logic within it can be unit tested independently and reused by multiple components. If possible, keep functions pure they are deterministic and very easy to unit test.


Thanks